remove assumptions about Python bytecode
Since the beginning of this project, the assumption that we are dealing
with

This means, the BasicBlock interface is now much simpler, with simply a
`label`, some `jump_targets` and `backedges` as attributes. Also,
`fallthrough` and `is_exiting` are now properties. This means,
`fallthrough` no longer needs to be handed in via every constructor call
and can be computed dynamically instead.

The Python payload is now handled by `PythonBytecodeLabel` and
`PythonBytecodeBlock`.