simplify simulator API
This simplifies the `run*` methods in the simulator such that only a
`Label` and not both `Label` and `BasicBlock` must be supplied. The
Simulator received a new attribute `region_stack` which memoizes the
current region and allows for a region aware `get_block` for any
simulator run. This significantly reduces code overhead and creates a
simple, easy to use API when iterating through the program.