disambiguate synth blocks
As the Label class is now removed we create types for all the synthetic
blocks. This ensures that tools that traverse the graph, like the
simulator and the rendering code to dispatch based on types of blocks
alone, without having to resort to doing partial matching on the block
names, which are strings.

The hope is that this will lead to more roboust code in the long-term.