fixup case of backedge from header to header
Fixes #45

Headers of multi-block loops that have a backedge to themselves, didn't
have this processed.

The case of:

`name not in doms[jt]`

accounts for the case, where there were multiple entry edges into the
loop and the original header has been replaced with a synthetic one.

An alternative fix would have been to simply delete this check, but then
this breaks the test case `test_double_header` (after hacking the
followup issues). This fix is non-intrusive and a test for the edge
cases has been added too.