fix missing variable assignment
Fixes #46

The problem was that, in case of a multi-header loop with a single exit,
the control-variable wasn't being setup correctly.

A further problem is that, even though this was tested, only the overall
block structure of the SCFG was being compared. However the instructions
of the blocks themselves were not, so this went unnoticed.

The fix for the test is to add the instructions contained within blocks
to the YAML and then we can test this properly.