update REAMDE and add makefile
The README now contains instructions based on the current state of the
available dependencies for Python 3.12. All of these are now available
from conda and no additional packages need to be installed via pip.

The makefiles is largely self-explanatory and contains all the low level
commands to build the project, run all the tests and to render the
documentation. It serves as a common entry-point for developers and
services alike.