enable `scfg.restructure()`
This moves the loop and branch restructuring code from the `ByteFlow`
class to the `SCFG` class. The `ByteFlow` class is scheduled for
eventual deletion anyway -- so this is a first step in that direction.

This is needed now, such that `SCFG` generated from Python abstract
syntax trees can also be transformed.