bugfix: continue from nested else
This test exposes a bug in a missing seal for using continue in the
nested for-else clause. Specifically, one needs to have an else-clause
with a series of linear statements and a continue to hit this. If you
try to break, the break index will be the same as the exit index for the
seal_inside_of_loop as the default_index is also the exit_index. So you
can only hit this with exit_index. In that case the block with the
continue does not point to the loop header but to the incorrect exit
instead.

The test has been constructed such that a potential reconstruction would
also fail execution with an incorrect return value.